refactor(web): migrate plugin toast calls to overlay ui toast
