+++ web/app/components/plugins/install-plugin/install-from-github/index.tsx
@@ -7,8 +7,8 @@ import * as React from 'react'
 import { useCallback, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import Modal from '@/app/components/base/modal'
-import Toast from '@/app/components/base/toast'
 import useGetIcon from '@/app/components/plugins/install-plugin/base/use-get-icon'
+import Toast from '@/app/components/plugins/utils/toast'
 import { cn } from '@/utils/classnames'
 import { InstallStepFromGitHub } from '../../types'
 import Installed from '../base/installed'
