+++ web/app/components/plugins/plugin-detail-panel/__tests__/detail-header.spec.tsx
@@ -2,7 +2,7 @@ import type { PluginDetail } from '../../types'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import * as amplitude from '@/app/components/base/amplitude'
-import Toast from '@/app/components/base/toast'
+import Toast from '@/app/components/plugins/utils/toast'
 import { PluginSource } from '../../types'
 import DetailHeader from '../detail-header'
