+++ web/app/components/plugins/plugin-detail-panel/endpoint-card.tsx
@@ -9,9 +9,9 @@ import ActionButton from '@/app/components/base/action-button'
 import Confirm from '@/app/components/base/confirm'
 import { CopyCheck } from '@/app/components/base/icons/src/vender/line/files'
 import Switch from '@/app/components/base/switch'
-import Toast from '@/app/components/base/toast'
 import Tooltip from '@/app/components/base/tooltip'
 import Indicator from '@/app/components/header/indicator'
+import Toast from '@/app/components/plugins/utils/toast'
 import { addDefaultValue, toolCredentialToFormSchemas } from '@/app/components/tools/utils/to-form-schema'
 import {
   useDeleteEndpoint,
