+++ web/app/components/plugins/plugin-detail-panel/endpoint-modal.tsx
@@ -8,8 +8,8 @@ import { useTranslation } from 'react-i18next'
 import ActionButton from '@/app/components/base/action-button'
 import Button from '@/app/components/base/button'
 import Drawer from '@/app/components/base/drawer'
-import Toast from '@/app/components/base/toast'
 import Form from '@/app/components/header/account-setting/model-provider-page/model-modal/Form'
+import Toast from '@/app/components/plugins/utils/toast'
 import { useRenderI18nObject } from '@/hooks/use-i18n'
 import { cn } from '@/utils/classnames'
 import { ReadmeEntrance } from '../readme-panel/entrance'
