+++ web/app/components/plugins/plugin-detail-panel/subscription-list/create/index.tsx
@@ -8,8 +8,8 @@ import { ActionButton, ActionButtonState } from '@/app/components/base/action-bu
 import Badge from '@/app/components/base/badge'
 import { Button } from '@/app/components/base/button'
 import CustomSelect from '@/app/components/base/select/custom'
-import Toast from '@/app/components/base/toast'
 import Tooltip from '@/app/components/base/tooltip'
+import Toast from '@/app/components/plugins/utils/toast'
 import { openOAuthPopup } from '@/hooks/use-oauth'
 import { useInitiateTriggerOAuth, useTriggerOAuthConfig, useTriggerProviderInfo } from '@/service/use-triggers'
 import { cn } from '@/utils/classnames'
