+++ web/app/components/plugins/plugin-detail-panel/subscription-list/edit/__tests__/apikey-edit-modal.spec.tsx
@@ -47,8 +47,8 @@ vi.mock('@/service/use-triggers', () => ({
   useTriggerPluginDynamicOptions: () => ({ data: [], isLoading: false }),
 }))
 
-vi.mock('@/app/components/base/toast', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('@/app/components/base/toast')>()
+vi.mock('@/app/components/plugins/utils/toast', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/app/components/plugins/utils/toast')>()
   return {
     ...actual,
     default: {
@@ -57,7 +57,7 @@ vi.mock('@/app/components/base/toast', async (importOriginal) => {
   }
 })
 
-vi.mock('@/app/components/base/toast/context', () => ({
+vi.mock('@/app/components/plugins/utils/toast/context', () => ({
   useToastContext: () => ({
     notify: (args: { type: string, message: string }) => mockToast(args),
     close: vi.fn(),
