+++ web/app/components/plugins/plugin-detail-panel/tool-selector/__tests__/index.spec.tsx
@@ -298,7 +298,7 @@ vi.mock('@/app/components/header/account-setting/model-provider-page/model-modal
 
 // Mock Toast - need to track notify calls for assertions
 const mockToastNotify = vi.fn()
-vi.mock('@/app/components/base/toast', () => ({
+vi.mock('@/app/components/plugins/utils/toast', () => ({
   default: { notify: (...args: unknown[]) => mockToastNotify(...args) },
 }))
 
@@ -1943,7 +1943,7 @@ describe('ToolCredentialsForm Component', () => {
       const saveBtn = screen.getByText(/save/i)
       fireEvent.click(saveBtn)
 
-      // Toast.notify should have been called with error (lines 49-50)
+      // notifyToast should have been called with error (lines 49-50)
       expect(mockToastNotify).toHaveBeenCalledWith(expect.objectContaining({ type: 'error' }))
       // onSaved should not be called because validation fails
       expect(onSaved).not.toHaveBeenCalled()
